+++ CHANGELOG.md
@@ -1,10 +1,12 @@
-## ADSBCOT beta
+## ADSBCOT 7.0.0
 
+- New for 2024!
 - Fixed formatting of CHANGELOG.md
+- Fixes #28: COT_STALE in documentation
 - Fixes #37: Move RTFD docs over to Markdown.
 - Fixes #38: Don't discard ImportError exception.
-
-
+- Fixes #39: Move CoT generation to PyTAK's `gen_cot_xml()` function.
+- Fixes #40: Add CoT `access` attribute.
 
 ## ADSBCOT 6.1.0